correct data type

+++ newImplement/Point.cpp
@@ -7,7 +7,7 @@ public:
     Point(){
 
     }
-    Point(int x,int y){
+    Point(double x,double y){
         this->x = x;
         this->y = y;
     }
